fix: Use int16_t (Temperature_t) for temperature binding
Change SensorBinding to use int16_t* instead of float* to match the
project's Temperature_t type (tenths of degrees).

Changes:
- Update SensorBinding struct: float* → int16_t*
- Add conversion in updateSensorReading(): value * 10.0f
- Update documentation to clarify Temperature_t format

This aligns with the project standard of using Temperature_t (int16_t)
representing tenths of degrees Celsius throughout the codebase.

+++ src/MB8ART.h
@@ -191,7 +191,7 @@ struct SensorHardwareConfig {
  * temperature and validity variables. It's initialized once at startup.
  */
 struct SensorBinding {
-    float* temperaturePtr;     // Pointer to temperature value in application
+    int16_t* temperaturePtr;   // Pointer to temperature in tenths of degrees (Temperature_t)
     bool* validityPtr;         // Pointer to validity flag in application
 };
